# ddtrace/tracer: optimize textmap.composeTracestate