--- a/ddtrace/tracer/spancontext_test.go
+++ b/ddtrace/tracer/spancontext_test.go
@@ -8,6 +8,7 @@ package tracer
 import (
 	"context"
 	"fmt"
+	"math"
 	"sync"
 	"testing"
 	"time"
@@ -854,4 +855,43 @@ func TestTraceIDEmpty(t *testing.T) {
 	tid := traceID([16]byte{})
 	tid[15] = 5
 	assert.False(t, tid.Empty())
+}
+
+func TestSpanIDHexEncoded(t *testing.T) {
+	sid := spanIDHexEncoded(5, 16)
+	assert.Equal(t, fmt.Sprintf("%016x", 5), sid)
+
+	sid = spanIDHexEncoded(5, 32)
+	assert.Equal(t, fmt.Sprintf("%032x", 5), sid)
+
+	sid = spanIDHexEncoded(math.MaxInt64, 68)
+	assert.Equal(t, fmt.Sprintf("%068x", math.MaxInt64), sid)
+
+	sid = spanIDHexEncoded(math.MaxInt64, 128)
+	assert.Equal(t, fmt.Sprintf("%0128x", math.MaxInt64), sid)
+}
+
+func BenchmarkSpanIDHexEncoded(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		_ = spanIDHexEncoded(32, 16)
+	}
+}
+
+func BenchmarkSpanIDSprintf(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		_ = fmt.Sprintf("%016x", 32)
+	}
+}
+
+func FuzzSpanIDHexEncoded(f *testing.F) {
+	f.Fuzz(func(t *testing.T, p int, v uint64) {
+		expected := fmt.Sprintf(
+			fmt.Sprintf("%%0%dx", p),
+			v,
+		)
+		actual := spanIDHexEncoded(v, p)
+		if actual != expected {
+			t.Fatalf("expected %s, got %s", expected, actual)
+		}
+	})
 }