--- a/ddtrace/tracer/textmap.go
+++ b/ddtrace/tracer/textmap.go
@@ -804,8 +804,17 @@ type stringMutator struct {
 	// be collapsed.
 	n bool
 	// fn is the function that implements the character replacement logic.
-	// It returns the rune to replace or drop (by returning -1), and a bool to tell if next consecutive
+	// It returns the rune to use as replacement and a bool to tell if next consecutive
 	// characters must be dropped if they fall in the currently matched character set.
+	// It's possible to return `-1` to immediately drop the current rune.
+	//
+	// This logic allows for:
+	// - Replace only the current rune: return <new value>, false
+	// - Drop only the current rune: return -1, false
+	// - Replace the current rune and drop the next consecutive runes if they match the same case: return <new value>, true
+	// - Drop all the consecutive runes matching the same case as the current one: return -1, true
+	//
+	// A known limitation is that we can only support a single case of consecutive runes.
 	fn func(rune) (rune, bool)
 }
 
@@ -844,7 +853,7 @@ func (sm *stringMutator) reset() {
 }
 
 var (
-	// keyRgx is used to sanitize the keys of the datadog propagating tags.
+	// keyDisallowedFn is used to sanitize the keys of the datadog propagating tags.
 	// Disallowed characters are comma (reserved as a list-member separator),
 	// equals (reserved for list-member key-value separator),
 	// space and characters outside the ASCII range 0x20 to 0x7E.
@@ -860,7 +869,7 @@ var (
 		return r, false
 	}
 
-	// valueRgx is used to sanitize the values of the datadog propagating tags.
+	// valueDisallowedFn is used to sanitize the values of the datadog propagating tags.
 	// Disallowed characters are comma (reserved as a list-member separator),
 	// semi-colon (reserved for separator between entries in the dd list-member),
 	// tilde (reserved, will represent 0x3D (equals) in the encoded tag value,
@@ -880,7 +889,7 @@ var (
 		return r, false
 	}
 
-	// originRgx is used to sanitize the value of the datadog origin tag.
+	// originDisallowedFn is used to sanitize the value of the datadog origin tag.
 	// Disallowed characters are comma (reserved as a list-member separator),
 	// semi-colon (reserved for separator between entries in the dd list-member),
 	// equals (reserved for list-member key-value separator),