### What does this PR do?

Replaces the usage of `fmt.Sprintf` and `regex` calls in `textmap.composeTracestate` to reduce allocations and CPU consumption.

### Motivation

This function has been reported as a significant contributor of tracer overhead.

Original benchmark results for the worst scenario:

```
BenchmarkComposeTracestate-10       448386          2429 ns/op         589 B/op         36 allocs/op
```

Benchmark results after removing all `fmt.Sprintf` usages:

```
BenchmarkComposeTracestate-10       508797          1976 ns/op         403 B/op         24 allocs/op
```

Benchmark results after replacing `regex` usages with tailored code:

```
BenchmarkComposeTracestate-10      1749620           675.3 ns/op       240 B/op         10 allocs/op
```

Benchmark results evolving the tailored code to a reusable state machine wrapped in a struct:

```
BenchmarkComposeTracestate-10    	 2625484	       444.6 ns/op	     160 B/op	       3 allocs/op
```

Benchmark results after replacing `fmt.Sprintf("%016x", spanID)` with a zero-allocation hand-coded hex encoding function:

```
BenchmarkComposeTracestate-10    	 3060886	       401.3 ns/op	     144 B/op	       2 allocs/op
```

### Reviewer's Checklist

- [ ] Changed code has unit tests for its functionality at or near 100% coverage.
- [ ] [System-Tests](https://github.com/DataDog/system-tests/) covering this feature have been added and enabled with the va.b.c-dev version tag.
- [ ] There is a benchmark for any new code, or changes to existing code.
- [ ] If this interacts with the agent in a new way, a system test has been added.
- [ ] Add an appropriate team label so this PR gets put in the right place for the release notes.
- [ ] Non-trivial go.mod changes, e.g. adding new modules, are reviewed by @DataDog/dd-trace-go-guild.


Unsure? Have a question? Request a review!